fix(admin-ui): move redux in domain layer
