+++ admin-ui/plugins/fido/domain/redux/sagas/FidoSaga.ts
@@ -13,9 +13,9 @@ import { getClient } from 'Redux/api/base'
 import { initAudit } from 'Redux/sagas/SagaUtils'
 import { updateToast } from 'Redux/features/toastSlice'
 import { postUserAction } from 'Redux/api/backend-api'
-import FidoApi from '../../api/FidoApi'
+import FidoApi from '../../../infrastructure/api/FidoApi'
 import { getFidoConfiguration, getFidoConfigurationResponse } from '../features/fidoSlice'
-import {  PayloadAction } from '@reduxjs/toolkit'
+import { PayloadAction } from '@reduxjs/toolkit'
 
 const JansConfigApi = require('jans_config_api')
 function* newFunction() {
