+++ admin-ui/plugins/fido/plugin-metadata.ts
@@ -1,5 +1,5 @@
-import { reducer as fidoReducer } from './infrastructure/redux/features/fidoSlice'
-import fidoSaga from './infrastructure/redux/sagas/FidoSaga'
+import { reducer as fidoReducer } from './domain/redux/features/fidoSlice'
+import fidoSaga from './domain/redux/sagas/FidoSaga'
 import { FIDO_READ, FIDO_WRITE } from '../../app/utils/PermChecker'
 import Fido from './presentation/pages/Fido'
